+++ packages/rx_bloc_cli/bin/compile.sh
@@ -11,5 +11,7 @@ dart run rx_bloc_cli create \
   --enable-social-logins=true \
   --enable-change-language=true \
   --enable-dev-menu=true \
+  --enable-patrol=true \
+
   example/test_app
 cp example/test_app/README.md example/