+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/bin/run_e2e_test_stg.sh
@@ -1,0 +1,2 @@
+#!/usr/bin/env sh
+patrol test --flavor staging --dart-define=flavor="staging" --no-uninstall